Update viewer_sequence.py
Sequence viewer fails if protocol directory structure does not exists. 
Therefore, check if "extra" directory exists before saving the sequence and if it does not exist then use /tmp.

When is this important?
1) open a new protocol that uses a sequence as input
2) select sequence
3) click eye to see sequence

viewer fails because it tries to create a file in the extra dir but that dir does not exist yet.